updates in case power goes out
