+++ email/emails/attendingEvent.handlebars
@@ -1,0 +1,29 @@
+<tr>
+    <td style="padding: 36px 30px 42px 30px">
+        <table role="presentation" style="width: 100%; border-collapse: collapse; border: 0; border-spacing: 0">
+            <tr>
+                <td style="padding: 0 0 36px 0; color: #353535">
+                    <h1 style="font-size: 24px; margin: 0 0 20px 0; font-family: Arial, sans-serif; text-align: center">
+                        You are attending -
+                        {{what}}!
+                    </h1>
+                    <p style="margin: 0 0 12px 0; font-size: 16px; line-height: 24px; font-family: Arial, sans-serif">
+                        {{what}}
+                        will be on
+                        {{date}}{{#if time}} at {{time}}.{{else}}.{{/if}}
+                    </p>
+                    {{#if Roommates}}
+                        <p style="margin: 0; font-size: 16px; line-height: 24px; font-family: Arial, sans-serif">
+                            Others attending are:
+                            <ul>
+                                {{#each Roommates}}
+                                    <li>{{first_name}} {{last_name}}</li>
+                                {{/each}}
+                            </ul>
+                        </p>
+                    {{/if}}
+                </td>
+            </tr>
+        </table>
+    </td>
+</tr>
