+++ email/emails/eventDeleted.handlebars
@@ -1,0 +1,14 @@
+<tr>
+    <td style="padding: 36px 30px 42px 30px">
+        <table role="presentation" style="width: 100%; border-collapse: collapse; border: 0; border-spacing: 0">
+            <tr>
+                <td style="padding: 0 0 36px 0; color: #353535">
+                    <h1 style="font-size: 24px; margin: 0 0 20px 0; font-family: Arial, sans-serif; text-align: center">
+                        {{what}}
+                        has been canceled. 😢
+                    </h1>
+                </td>
+            </tr>
+        </table>
+    </td>
+</tr>
