+++ package.json
@@ -27,6 +27,7 @@
     "express-session": "^1.17.3",
     "mysql2": "^3.1.1",
     "nodemailer": "^6.9.1",
+    "nodemailer-express-handlebars": "^6.0.0",
     "path": "^0.12.7",
     "sequelize": "^6.28.0"
   }
